+++ BackEnd/Genesis/src/main/resources/application.properties
@@ -10,8 +10,8 @@ spring.datasource.username=GENESIS
 spring.datasource.password=GENESIS
 
 # defini\u00e7\u00f5es do hibernate
-spring.jpa.hibernate.ddl-auto=update
-#spring.jpa.hibernate.ddl-auto=create-drop
+#spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create-drop
 spring.datasource.initialize=false
 spring.jpa.hibernate.show_sql=true
 spring.jpa.hibernate.format_sql=true
